calling move from rom_nom module

+++ move.py
@@ -1,11 +1,11 @@
 import os, shutil, zipfile, time
 
 def move_roms(move_from, move_to):
-    print('starting move...\n')
+    print(move_from, move_to)
+    print('\nstarting move...\n')
     start = time.perf_counter()
     os.chdir(move_from)
     cwd = os.getcwd()
-    dst_dir = (move_to)
     os.makedirs('zip_files', exist_ok=True)
 
     if not os.path.isdir(move_to):
@@ -22,13 +22,13 @@ def move_roms(move_from, move_to):
 
         os.remove(zip_file)
 
-    print('copying files to flash drive...\n')
-    for item in os.listdir(cwd):
-        if item.endswith('nes'):
-            shutil.copy(item, dst_dir)
-        elif item == 'zip_files' or item == 'inventory.txt':
-            continue
-        os.remove(item)
+    if move_from != move_to:
+        print(fr'copying files to {move_to}...\n')
+        for item in os.listdir(cwd):
+            if item == 'zip_files' or item == 'inventory.txt' or os.path.isdir(item):
+                continue
+            shutil.copy(item, move_to)
+            os.remove(item)
 
     end = time.perf_counter()
     print(f'move completed in {end - start:0.2f} seconds\n')
